requirements.txt
django==4.2.2
psycopg2-binary==2.9.9
pillow==10.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-yasg==1.21.7
setuptools==70.3.0
requests==2.32.3
celery==5.4.0
redis==5.0.7
gevent==24.2.1
django-celery-beat==2.6.0
coverage==7.5.4
